[Refactor]: GithubClient 인터페이스 분리 (#25)
