--- src/main/java/team/gdsc/code2cv/feature/auth/repository/GithubOAuth2Client.java
@@ -1,4 +1,0 @@
-package team.gdsc.code2cv.feature.auth.repository;
-
-public class GithubOAuth2Client {
-}
